separate cli into file
