+++ scrmbl/__init__.py
@@ -1,3 +1,5 @@
-from scrmbl.scrmbl import echo, __version__
+from scrmbl.scrmbl import echo
 
 __all__ = ['echo']
+
+__version__ = '0.1.1'
